switchbot-logger/pyproject.toml
[tool.poetry]
name = "switchbot-logger"
version = "0.1.0"
description = "Logger for Switchbot sensors"
authors = ["yadach <47210083+yadach@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "switchbot_logger"}]

[tool.poetry.dependencies]
python = "^3.9"
bluepy = "^1.3.0"
pyyaml = "^6.0.1"
mysql-connector-python = "^8.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.1.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = ".venv"
ignore = "H102,H103,W503,E203"
max-line-length = 100

[tool.black]
line-length = 100
exclude = '''
(
    .pytest_cache
    | .venv
)
'''

[tool.isort]
include_trailing_comma = true
line_length = 100
sort_order = "native"
force_single_line = true
lines_between_types = 0
